+++ TYPO3.Flow/Classes/TYPO3/Flow/Property/PropertyMappingConfigurationBuilder.php
@@ -17,6 +17,7 @@ use TYPO3\Flow\Annotations as Flow;
  * This builder creates the default configuration for Property Mapping, if no configuration has been passed to the Property Mapper.
  *
  * @Flow\Scope("singleton")
+ * @deprecated Will be removed in next major Flow version.
  */
 class PropertyMappingConfigurationBuilder
 {
@@ -25,6 +26,8 @@ class PropertyMappingConfigurationBuilder
      *
      * @param string $type the implementation class name of the PropertyMappingConfiguration to instantiate; must be a subclass of TYPO3\Flow\Property\PropertyMappingConfiguration
      * @return \TYPO3\Flow\Property\PropertyMappingConfiguration
+     * @deprecated fully replace by PropertyMapper::buildPropertyMappingConfiguration and removed in next major Flow version
+     * @see PropertyMapper::buildPropertyMappingConfiguration
      */
     public function build($type = \TYPO3\Flow\Property\PropertyMappingConfiguration::class)
     {
